refactor(ui): share floating-layer dismiss guard, apply to Sheet
Extract the dialog's nested-floating-layer dismiss guard (target allowlist +
capture-phase open-state flag) into a reusable useFloatingLayerDismissGuard hook
and consume it from DialogContent. Apply the same guard to SheetContent, which
is built on the Dialog primitive and shares the close-on-outside behaviour, so a
future Select/dropdown inside a Sheet will not close the sheet.

AlertDialog is intentionally left as-is: Radix AlertDialog is non-dismissible on
outside pointer interaction by design (it omits those handlers), so it cannot hit
this bug.
